+++ SkrisForum.Data/SkrisForumDBContext.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.EntityFrameworkCore;
-using SkrisForum.Data.Entity;
+using SkrisForum.Data.Entities;
 
 namespace SkrisForum.Data
 {
